BHA: Amend task and flow decoration and add dates
This changes the previous subflows to be tasks and the previous tasks to be general functions to better reflect their status and allow for configuration.
Tasks have now been tagged BHA.

The last update date is now calculated rather than hardcoded.

raise_for_status() has been applied to the response in the download_file function