Rapid: Create rapid_horseracing_fetcher flow and tasks
This creates a flow to (initially) fetch rapid horseracing racecards via the API.
It makes use of a modified download_file helper method which now takes headers and params
and an additional get_files helper, used when calculating the next date to be fetched.

It loops through and fetches cards according to the rate limits set as constants.